Introduce a global helper file for general helpers
