+++ components/shared/workCard/helper.ts
@@ -1,11 +1,11 @@
 import { flatten } from "lodash"
 
 import { Cover } from "@/lib/cover-service-api/model"
-import { filterFalsyValuesFromArray } from "@/lib/fetchers/helper"
 import { SearchWithPaginationQuery, WorkTeaserFragment } from "@/lib/graphql/generated/fbi/graphql"
 
 import { CoverImageUrls } from "./../../../lib/cover-service-api/model/coverImageUrls"
 import { GeneralMaterialTypeCode } from "./../../../lib/graphql/generated/fbi/graphql"
+import { filterFalsyValuesFromArray } from "@/lib/helper"
 
 export const displayCreators = (
   creators: SearchWithPaginationQuery["search"]["works"][0]["creators"],
